Enhance: Restyle the node handles
The initial handles for the resize and connector points were, not to put too fine a point on it, ugly. They lacked a certain level of refinement. I have updated them to make the connector points round and to change the stroke and background colours to something that is visually more appealing (to me anyway).